fix(ops): wire /api/ops/clickhouse/explain into the Hono router
#4434 placed the explain handler at src/pages/api/ops/clickhouse/
explain.ts, but the langwatch app has been Hono-based for a while now:
files under src/pages/api/ are NOT auto-routed, routes are explicitly
registered in src/server/api-router.ts. The handler kept compiling, the
unit + "integration" tests kept passing (they invoked the function
directly), and CI was green — but prod returned 404 for every request
because the path was never bound to HTTP.

Caught by trying to actually call the endpoint from the agent box.

Changes:
- Extract pure helpers (buildExplainQuery, stripCommentsAndStrings,
  redactQueryForAudit, getOpsClickHouseClient, ALLOWED_EXPLAIN_TYPES,
  CLICKHOUSE_GUARDRAILS, explainBodySchema, _resetOpsClickHouseClient
  ForTesting) into src/server/ops/explain-core.ts so the route file
  stays focused on HTTP plumbing.
- New Hono route at src/server/routes/ops.ts. POST
  /api/ops/clickhouse/explain. Same defenses-in-depth as before (Bearer
  LANGWATCH_OPS_API_KEY constant-time, query normalisation + regex
  filter, EXPLAIN wrapping + type allowlist, dedicated langwatch_ops
  CH client via CLICKHOUSE_OPS_URL with prod fail-closed, per-query
  guardrails, audit-redacted log). Auth is handlerManagedAuth — the
  policy comment documents the contract.
- Mount in src/server/api-router.ts alongside misc/sse/etc.
- Delete the dead src/pages/api/ops/clickhouse/* entirely (handler +
  old tests).

The lesson — direct handler invocation isn't an integration test. The
new src/server/routes/__tests__/ops-explain.integration.test.ts
imports createApiRouter() (the FULL router, not just the ops module)
and exercises every case via router.request(path, {...}) so a missing
api.route("/", opsApp) in api-router.ts shows up as a 404 here instead
of green CI + 404 in prod. 12 HTTP cases covered: 200 happy path,
cross-tenant, PLAN vs PIPELINE diff, 400 on multi-statement / table
function / string-vs-comment bypass / system.* / missing key, 401 on
wrong key, 405-or-404 on GET, 502 on missing table, realistic
arrayJoin shape. 28 unit cases for the pure helpers in
src/server/ops/__tests__/explain-core.unit.test.ts. All green.
