+++ langwatch/src/server/api-router.ts
@@ -55,6 +55,7 @@ import { app as cronApp } from "./routes/cron";
 import { app as evaluationsLegacyApp } from "./routes/evaluations-legacy";
 import { app as healthApp } from "./routes/health";
 import { app as miscApp } from "./routes/misc";
+import { app as opsApp } from "./routes/ops";
 import { app as sseApp } from "./routes/sse";
 import { app as tracesLegacyApp } from "./routes/traces-legacy";
 import { app as trpcApp } from "./routes/trpc";
@@ -154,6 +155,7 @@ export function createApiRouter() {
   api.route("/", evaluationsLegacyApp);
   api.route("/", healthApp);
   api.route("/", miscApp);
+  api.route("/", opsApp);
   api.route("/", sseApp);
   api.route("/", tracesLegacyApp);
   api.route("/", trpcApp);
